requirements.txt
delta-spark==3.2.0
jupyter
pandas==1.5.3
numpy
pyspark==3.5.1